removed createoptionsmenu from VP tab fragments, homefragment controls optionsmenu inflation
