+++ WhatsApp/app/src/main/java/com/example/whatsapp/ui/fragments/contactsselect/ContactsSelectFragment.kt
@@ -143,20 +143,30 @@ class ContactsSelectFragment : BaseFragment<FragmentContactsSelectBinding>(), Ac
     //region Options Menu
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         Log.i("ACTIONTEST", "OptionsMenu Created")
+        activity?.invalidateOptionsMenu()
         inflater.inflate(R.menu.contacts_menu, menu)
         super.onCreateOptionsMenu(menu, inflater)
     }
-
-    override fun onPrepareOptionsMenu(menu: Menu) {
-        super.onPrepareOptionsMenu(menu)
-        menu.findItem(R.id.searchFragment).isVisible = false
-        menu.findItem(R.id.itemLinkedDevices).isVisible = false
-        menu.findItem(R.id.itemStarredMessages).isVisible = false
-        menu.findItem(R.id.itemSettings).isVisible = false
-        menu.findItem(R.id.itemNewGroup).isVisible = false
-        menu.findItem(R.id.itemNewBroadcast).isVisible = false
-    }
-
+//    override fun onPrepareOptionsMenu(menu: Menu) {
+//        super.onPrepareOptionsMenu(menu)
+//        val itemList = listOf(
+//            R.id.searchFragment,
+//            R.id.itemLinkedDevices,
+//            R.id.itemStarredMessages,
+//            R.id.itemSettings,
+//            R.id.itemNewGroup,
+//            R.id.itemNewBroadcast,
+//            R.id.clearCallLog,
+//            R.id.settings,
+//            R.id.search)
+//        for (item in itemList) {
+//            if (menu.findItem(item) == null) {
+//                continue
+//            } else {
+//                menu.findItem(item).isVisible = false
+//            }
+//        }
+//    }
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return when (item.itemId) {
             R.id.search -> {
