+++ WhatsApp/app/src/main/java/com/example/whatsapp/ui/fragments/home/HomeFragment.kt
@@ -2,6 +2,8 @@ package com.example.whatsapp.ui.fragments.home
 
 import android.util.Log
 import android.view.LayoutInflater
+import android.view.Menu
+import android.view.MenuInflater
 import android.view.View
 import androidx.appcompat.content.res.AppCompatResources
 import androidx.core.content.res.ResourcesCompat
@@ -30,6 +32,7 @@ import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class HomeFragment : BaseFragment<FragmentHomeBinding>() {
+    private val TAG = "HOMETEST"
 
     //region Variables
     private val viewModel by viewModels<HomeFragmentVM>()
@@ -46,6 +49,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>() {
 
     override fun initViews() {
         super.initViews()
+        setHasOptionsMenu(true)
         //region Tabs Initialisation
         val fragments = listOf(
             CameraFragment.newInstance(),
@@ -136,6 +140,21 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>() {
                 }
             }
         }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        Log.i(TAG, "TAB POSITION $currentTabPosition")
+        if (currentTabPosition == CHATS_FRAGMENT_INDEX) {
+            inflater.inflate(R.menu.home_chats_menu, menu)
+        } else if (currentTabPosition == STATUS_FRAGMENT_INDEX) {
+            inflater.inflate(R.menu.home_status_menu, menu)
+        } else if (currentTabPosition == CALLS_FRAGMENT_INDEX) {
+            inflater.inflate(R.menu.home_calls_menu, menu)
+        } else {
+            activity?.invalidateOptionsMenu()
+            menu.clear()
+        }
+        super.onCreateOptionsMenu(menu, inflater)
     }
 
     override fun observeViewModel() {
